fix : unimplemented method 제거

+++ src/main/java/com/moru/backend/domain/auth/application/LogoutService.java
@@ -2,6 +2,8 @@ package com.moru.backend.domain.auth.application;
 
 import com.moru.backend.domain.auth.application.TokenBlacklistService;
 import com.moru.backend.domain.user.domain.User;
+import com.moru.backend.global.exception.CustomException;
+import com.moru.backend.global.exception.ErrorCode;
 import com.moru.backend.global.jwt.JwtProvider;
 import com.moru.backend.global.redis.RefreshTokenRepository;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,9 +20,14 @@ public class LogoutService {
     private final TokenBlacklistService tokenBlacklistService;
 
     public void logout(User user, String accessToken) {
-        // 저장된 refresh Token 삭제
-        refreshTokenRepository.delete(user.getId().toString()); 
-        // access token 블랙리스트에 저장
+        // // 저장된 refresh Token 삭제
+        // refreshTokenRepository.delete(user.getId().toString()); 
+        // // access token 블랙리스트에 저장
+        // tokenBlacklistService.addBlacklist(accessToken);
+        if (user == null || user.getId() == null) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED);
+        }
+        refreshTokenRepository.delete(user.getId().toString());
         tokenBlacklistService.addBlacklist(accessToken);
     }
 }
